+++ src/components/muiComponents/imageAvatars.jsx
@@ -1,13 +1,16 @@
 import * as React from 'react';
 import Avatar from '@mui/material/Avatar';
 import Stack from '@mui/material/Stack';
+import { Link } from 'react-router-dom';
 export default function ImageAvatars() {
   return (
     <Stack sx={{ display: 'flex ', justifyContent: 'space-around' }} direction="row" spacing={2}>
-      <Avatar
-        alt="Remy Sharp"
-        src="https://logo-logos.com/wp-content/uploads/2016/10/Instagram_logo_icon.png"
-      />
+      <Link to="www.instagram.com">
+        <Avatar
+          alt="Remy Sharp"
+          src="https://logo-logos.com/wp-content/uploads/2016/10/Instagram_logo_icon.png"
+        />
+      </Link>
       <Avatar
         alt="Travis Howard"
         src="https://upload.wikimedia.org/wikipedia/commons/thumb/5/51/Facebook_f_logo_%282019%29.svg/2048px-Facebook_f_logo_%282019%29.svg.png"
